requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.2
celery==5.4.0
redis==5.0.8
scikit-learn==1.5.1
pandas==2.2.2
joblib==1.4.2
yt-dlp==2025.01.15   # пример, можно взять актуальную стабильную
openai-whisper==20231117
